+++ suite-common/wallet-core/src/send/sendFormThunks.ts
@@ -73,15 +73,15 @@ import {
     signSolanaSendFormTransactionThunk,
 } from './sendFormSolanaThunks';
 import {
-    composeTronTransactionFeeLevelsThunk,
-    signTronSendFormTransactionThunk,
-} from './sendFormTronThunks';
-import {
     type ComposeFeeLevelsError,
     type PushTransactionError,
     type SignTransactionError,
     type SignTransactionTimeoutError,
 } from './sendFormTypes';
+import {
+    composeTronTransactionFeeLevelsThunk,
+    signTronSendFormTransactionThunk,
+} from './tron/sendFormTronThunks';
 import { accountsActions } from '../accounts/accountsActions';
 import { selectAccountByKey } from '../accounts/accountsSelectors';
 import { syncAccountsWithBlockchainThunk } from '../blockchain/blockchainThunks';
