+++ suite-common/wallet-core/src/send/tron/types.ts
@@ -1,0 +1,3 @@
+import { type BlockchainLinkResponse } from '@trezor/blockchain-link';
+
+export type EstimateFeeLevel = BlockchainLinkResponse<'estimateFee'>[number];
